gemini_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.7.2  # El SDK moderno
pydantic==2.9.2
python-dotenv==1.0.1
google-api-core==2.19.0